empty database

+++ backBatiUni/modelData/buildDataBase.py
@@ -218,10 +218,6 @@ class CreateNewDataBase:
           os.rmdir(os.path.join(newPath))
         else:
           os.remove(os.path.join(dir, file))
-    for table in CreateNewDataBase.listTable.values():
-      table.objects.all().delete()
-      tableName = table.objects.model._meta.db_table
-      self.cursor.execute(f"ALTER TABLE {tableName} AUTO_INCREMENT=1;")
     for user in User.objects.all():
       if user.username != "jlw":
         userProfileList = UserProfile.objects.filter(userNameInternal = user)
@@ -229,6 +225,12 @@ class CreateNewDataBase:
           userProfile = userProfileList[0]
           company = Company.objects.get(id = userProfile.Company.id)
           stripe.Customer.delete(company.stripeCustomerId)
+    for table in CreateNewDataBase.listTable.values():
+      table.objects.all().delete()
+      tableName = table.objects.model._meta.db_table
+      self.cursor.execute(f"ALTER TABLE {tableName} AUTO_INCREMENT=1;")
+    for user in User.objects.all():
+      if user.username != "jlw":
         user.delete()
     self.cursor.execute("ALTER TABLE auth_user AUTO_INCREMENT=1;")
     return {"emptyDataBase":"OK"}
